Move formation templates into aws and os directories
