+++ lib/cfndsl/aws/cloud_formation_template.rb
@@ -1,0 +1,15 @@
+require 'cfndsl/orchestration_template'
+
+module CfnDsl
+  class CloudFormationTemplate < OrchestrationTemplate
+    def self.template_types
+      CfnDsl::AWS::Types::Types_Internal
+    end
+
+    def self.type_module
+      CfnDsl::AWS::Types
+    end
+
+    create_types
+  end
+end
